refactor ruby scripts to a module and class
